refactor the index.js
Move the BotManager to a separate file and then called it in the index.js
